requirements.txt
aiocontextvars==0.2.2
aioredis==2.0.1
aiosmtplib==1.1.6
anyio==3.5.0
asgiref==3.4.1
async-generator==1.10
async-timeout==4.0.2
asyncpg==0.25.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.11
click==8.0.3
contextlib2==21.6.0
contextvars==2.4
# cryptography==36.0.1
databases==0.5.5
dataclasses==0.8
Deprecated==1.2.13
dnspython==2.2.0
email-validator==1.1.3
encode==0.0.0
fakeredis==1.7.1
fastapi==0.73.0
fastapi-mail==1.0.4
greenlet==1.1.2
h11==0.12.0
h2==3.2.0
hpack==3.0.0
httpcore==0.13.7
httptools==0.2.0
httpx==0.19.0
hyperframe==5.2.0
idna==2.10
immutables==0.16
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
# orjson==3.6.1
packaging==21.3
psycopg2==2.9.3
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
pyparsing==3.0.7
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==5.4.1
redis==4.1.4
requests==2.27.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.31
starlette==0.17.1
typing_extensions==4.0.1
ujson==4.3.0
urllib3==1.26.8
uvicorn==0.15.0
uvloop==0.14.0
watchgod==0.7
websockets==9.1
wrapt==1.13.3
zipp==3.6.0
